add support for ExpressionType.TypeEqual
TypeIs is not the only kind of TypeBinaryExpression node. TypeIs represents the c# "is" or vb "TypeOf..Is" operator, but there is also a TypeEqual node which represents an exact type equality check. foo TypeEqual Foo iff foo is an instance of Foo, not of subclasses.

There is no C# syntax which produces a TypeEqual node as far as I'm aware, but it works fine with linq2objects and has important use cases in a linq provider I maintain. This pull request updates TypeBinaryExpressionNode.cs to support deserialising into a TypeEqual node.